% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRaschBayes-package.R
\docType{package}
\name{easyRaschBayes-package}
\alias{easyRaschBayes}
\alias{easyRaschBayes-package}
\title{easyRaschBayes: Bayesian Rasch Analysis Using 'brms'}
\description{
Reproduces classic Rasch psychometric analysis features using Bayesian item response theory models fitted with 'brms' following Bürkner (2021) \doi{10.18637/jss.v100.i05} and Bürkner (2020) \doi{10.3390/jintelligence8010005}. Supports both dichotomous and polytomous Rasch models. Features include posterior predictive item fit, conditional infit, item-restscore associations, person fit, differential item functioning, local dependence assessment via Q3 residual correlations, dimensionality assessment with residual principal components analysis, person-item targeting plots, item category probability curves, and reliability using relative measurement uncertainty following Bignardi et al. (2025) \doi{10.31234/osf.io/h54k8_v1}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pgmj/easyRaschBayes}
  \item \url{https://pgmj.github.io/easyRaschBayes/}
  \item Report bugs at \url{https://github.com/pgmj/easyRaschBayes/issues}
}

}
\author{
\strong{Maintainer}: Magnus Johansson \email{pgmj@pm.me} (\href{https://orcid.org/0000-0003-1669-592X}{ORCID})

Other contributors:
\itemize{
  \item Giacomo Bignardi (\href{https://orcid.org/0000-0002-1153-0838}{ORCID}) (RMU reliability code) [contributor]
  \item Kristoffer Magnusson (\href{https://orcid.org/0000-0003-0713-0556}{ORCID}) (hurdle_acat custom brms family) [contributor]
}

}
\keyword{internal}
