% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_hpcm.R
\name{hurdle_acat_stanvars}
\alias{hurdle_acat_stanvars}
\title{Stan Function Block for \code{hurdle_acat}}
\usage{
hurdle_acat_stanvars()
}
\value{
A \code{stanvars} object adding the
\code{hurdle_acat_merged_lpmf} (and helper \code{acat_logit_lpmf})
to the \code{functions} block of the generated Stan model.
}
\description{
Returns a \code{\link[brms]{stanvar}} containing the Stan
implementation of the hurdle-acat log-PMF used by
\code{\link{hurdle_acat}}. Pass it as the \code{stanvars} argument
to \code{\link[brms]{brm}}.
}
\seealso{
\code{\link{hurdle_acat}}
}
\author{
Kristoffer Magnusson
}
