% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_hpcm.R
\name{log_lik_hurdle_acat}
\alias{log_lik_hurdle_acat}
\title{Log-Likelihood Method for \code{hurdle_acat}}
\usage{
log_lik_hurdle_acat(i, prep)
}
\arguments{
\item{i}{Observation index (supplied by brms).}

\item{prep}{A brms prep object (supplied by brms).}
}
\value{
A numeric vector of length \code{prep$ndraws}.
}
\description{
Not called directly. brms invokes this when \code{\link[brms]{log_lik}}
is called on a model fitted with \code{\link{hurdle_acat}}.
}
\author{
Kristoffer Magnusson
}
\keyword{internal}
