% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_hpcm.R
\name{posterior_predict_hurdle_acat}
\alias{posterior_predict_hurdle_acat}
\title{Posterior-Predict Method for \code{hurdle_acat}}
\usage{
posterior_predict_hurdle_acat(i, prep, ...)
}
\arguments{
\item{i}{Observation index (supplied by brms).}

\item{prep}{A brms prep object (supplied by brms).}

\item{...}{Currently unused.}
}
\value{
An integer vector of length \code{prep$ndraws} containing
replicated category labels in \eqn{0, 1, \ldots, K - 1}.
}
\description{
Not called directly. brms invokes this when
\code{\link[brms]{posterior_predict}} is called on a model fitted
with \code{\link{hurdle_acat}}.
}
\author{
Kristoffer Magnusson
}
\keyword{internal}
