% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRaschBayes-package.R
\docType{package}
\name{easyRaschBayes-package}
\alias{easyRaschBayes}
\alias{easyRaschBayes-package}
\title{easyRaschBayes: Bayesian Rasch Analysis Using 'brms'}
\description{
Provides functions to reproduce classic Rasch analysis features using Bayesian item response theory (IRT) models fitted with 'brms'. Supports both dichotomous Rasch models and polytomous partial credit or rating scale models. Features include posterior predictive item fit, person fit, differential item functioning (DIF), local dependence assessment via Q3 residual correlations, dimensionality assessment via residual principal components analysis, person-item targeting plots, item category probability curves, and reliability via Relative Measurement Uncertainty.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pgmj/easyRaschBayes}
  \item Report bugs at \url{https://github.com/pgmj/easyRaschBayes/issues}
}

}
\author{
\strong{Maintainer}: Magnus Johansson \email{pgmj@pm.me} (\href{https://orcid.org/0000-0003-1669-592X}{ORCID})

Other contributors:
\itemize{
  \item Giacomo Bignardi (RMU reliability code) [contributor]
}

}
\keyword{internal}
